pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc2graph"
version = "2.0.0"
description = "Estrae knowledge graph da documenti via LLM e genera un visualizzatore HTML interattivo"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["knowledge-graph", "llm", "nlp", "neo4j", "visualization"]

dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
]

[project.optional-dependencies]
full = [
    "rapidfuzz>=3.0.0",
    "tiktoken>=0.6.0",
    "rich>=13.0.0",
    "jinja2>=3.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
    "black",
    "ruff",
]

[project.scripts]
doc2graph = "doc2graph.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["doc2graph*"]

[tool.setuptools.package-data]
doc2graph = ["../templates/*.jinja2"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
